+++ gui/src/components/BalancesList.tsx
@@ -53,7 +53,7 @@ function BalancesERC20() {
   return (
     <>
       {filteredBalances.map(({ contract, balance, metadata }) => (
-        <BalanceERC20
+        <BalanceItem
           key={contract}
           contract={contract}
           balance={BigInt(balance)}
@@ -62,29 +62,6 @@ function BalancesERC20() {
         />
       ))}
     </>
-  );
-}
-
-function BalanceERC20({
-  contract,
-  balance,
-  decimals,
-  symbol,
-}: {
-  contract: Address;
-  balance: bigint;
-  decimals: number;
-  symbol: string;
-}) {
-  if (!symbol || !decimals) return null;
-
-  return (
-    <BalanceItem
-      contract={contract}
-      balance={balance}
-      decimals={decimals}
-      symbol={symbol}
-    />
   );
 }
 
@@ -106,6 +83,9 @@ function BalanceItem({
   const truncatedBalance =
     balance - (balance % BigInt(Math.ceil(minimum * 10 ** decimals)));
 
+  if (!symbol || !decimals) return null;
+  console.log(contract);
+
   return (
     <ListItem>
       <ListItemAvatar>
@@ -115,7 +95,7 @@ function BalanceItem({
       </ListItemAvatar>
       <ListItemText
         secondary={`${symbol} ${
-          contract && `(${truncateEthAddress(contract)})`
+          contract ? `(${truncateEthAddress(contract)})` : ``
         }`}
       >
         <CopyToClipboard label={balance.toString()}>
