+++ gui/src/components/Settings/Wallet/Impersonator.tsx
@@ -1,0 +1,3 @@
+export function ImpersonatorForm() {
+  return <></>;
+}
