+++ gui/src/components/index.ts
@@ -19,10 +19,7 @@ export { Peers } from "./Peers";
 export { QuickAddressSelect } from "./QuickAddressSelect";
 export { QuickNetworkSelect } from "./QuickNetworkSelect";
 export { QuickWalletSelect } from "./QuickWalletSelect";
-export { Settings } from "./Settings";
-export { SettingsGeneral } from "./SettingsGeneral";
-export { SettingsNetwork } from "./SettingsNetwork";
-export { SettingsWallets } from "./SettingsWallets";
+export { Settings } from "./Settings/";
 export { TabPanel } from "./TabPanel";
 export { TxReviewDialog } from "./TxReviewDialog";
 export { Txs } from "./Txs";
